Revert "LPS-70805 Change to Integer"
This reverts commit 754d486d37dac3d2890b0095e98831c4e9b5a1b5.

+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-form-evaluator-impl/src/main/java/com/liferay/dynamic/data/mapping/form/evaluator/impl/internal/functions/JumpPageFunction.java
@@ -33,10 +33,10 @@ public class JumpPageFunction implements DDMExpressionFunction {
 			throw new IllegalArgumentException("Two parameters are expected");
 		}
 
-		Integer fromPageIndex = (Integer)parameters[0];
-		Integer toPageIndex = (Integer)parameters[1];
+		Double fromPageIndex = (Double)parameters[0];
+		Double toPageIndex = (Double)parameters[1];
 
-		_pageFlow.put(fromPageIndex, toPageIndex);
+		_pageFlow.put(fromPageIndex.intValue(), toPageIndex.intValue());
 
 		return true;
 	}
